feat(ingest): remove need for sink block in UI based ingestion (#5208)
