+++ datahub-web-react/src/app/ingest/source/builder/NameSourceStep.tsx
@@ -83,7 +83,7 @@ export const NameSourceStep = ({ state, updateState, prev, submit }: StepProps) 
                                 Advanced: Provide a custom CLI version to use for ingestion.
                             </Typography.Paragraph>
                             <Input
-                                placeholder="0.8.38"
+                                placeholder="0.8.38.2"
                                 value={state.config?.version || ''}
                                 onChange={(event) => setVersion(event.target.value)}
                             />
