+++ datahub-web-react/src/app/ingest/source/conf/azure/azure.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import azureLogo from '../../../../../images/azure-ad.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: azure-ad
@@ -29,12 +27,7 @@ source:
         # users_pattern:
         #    allow:
         #        - ".*"
-sink: 
-    type: datahub-rest
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const azureAdConfig: SourceConfig = {
     type: 'azure-ad',
