+++ datahub-web-react/src/app/ingest/source/conf/bigquery/bigquery.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import bigqueryLogo from '../../../../../images/bigquerylogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: bigquery
@@ -21,12 +19,7 @@ source:
 
             client_email: # Your BQ client email, e.g. "test@suppproject-id-1234567.iam.gserviceaccount.com"
             client_id: # Your BQ client id, e.g. "123456678890"
-sink: 
-    type: datahub-rest
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const bigqueryConfig: SourceConfig = {
     type: 'bigquery',
