+++ datahub-web-react/src/app/ingest/source/conf/glue/glue.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import glueLogo from '../../../../../images/gluelogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: glue
@@ -24,12 +22,7 @@ source:
         # table_pattern:
         #    allow:
         #        - "avro"
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const glueConfig: SourceConfig = {
     type: 'glue',
