+++ datahub-web-react/src/app/ingest/source/conf/hive/hive.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import hiveLogo from '../../../../../images/hivelogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source: 
     type: hive
@@ -15,13 +13,7 @@ source:
         # Add secret in Secrets Tab with relevant names for each variable
         username: "\${HIVE_USERNAME}" # Your Hive username, e.g. admin
         password: "\${HIVE_PASSWORD}"# Your Hive password, e.g. password_01
-
-sink: 
-    type: datahub-rest
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const hiveConfig: SourceConfig = {
     type: 'hive',
