+++ datahub-web-react/src/app/ingest/source/conf/kafka/kafka.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import kafkaLogo from '../../../../../images/kafkalogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: kafka
@@ -23,12 +21,7 @@ source:
             # Uncomment and add secrets in Secrets Tab
             # schema_registry_config:
             #     basic.auth.user.info: "\${REGISTRY_API_KEY_ID}:\${REGISTRY_API_KEY_SECRET}"
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const kafkaConfig: SourceConfig = {
     type: 'kafka',
