+++ datahub-web-react/src/app/ingest/source/conf/looker/looker.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import lookerLogo from '../../../../../images/lookerlogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: looker
@@ -14,12 +12,7 @@ source:
         # Add secret in Secrets Tab with relevant names for each variable
         client_id: "\${LOOKER_CLIENT_ID}" # Your Looker client id, e.g. admin
         client_secret: "\${LOOKER_CLIENT_SECRET}" # Your Looker password, e.g. password_01
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const lookerConfig: SourceConfig = {
     type: 'looker',
