+++ datahub-web-react/src/app/ingest/source/conf/mysql/mysql.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import mysqlLogo from '../../../../../images/mysqllogo-2.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source: 
     type: mysql
@@ -23,12 +21,7 @@ source:
         # Profiling
         profiling:
             enabled: false
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const mysqlConfig: SourceConfig = {
     type: 'mysql',
