+++ datahub-web-react/src/app/ingest/source/conf/okta/okta.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import oktaLogo from '../../../../../images/oktalogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source:
     type: okta
@@ -31,12 +29,7 @@ source:
         # Optional: Include deprovisioned or suspended Okta users in the ingestion.
         # include_deprovisioned_users = False
         # include_suspended_users = False
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const oktaConfig: SourceConfig = {
     type: 'okta',
