+++ datahub-web-react/src/app/ingest/source/conf/redshift/redshift.ts
@@ -1,8 +1,6 @@
 import { SourceConfig } from '../types';
 import redshiftLogo from '../../../../../images/redshiftlogo.png';
 
-const baseUrl = window.location.origin;
-
 const placeholderRecipe = `\
 source: 
     type: redshift
@@ -23,12 +21,7 @@ source:
         # Profiling
         profiling:
             enabled: false
-sink: 
-    type: datahub-rest 
-    config: 
-        server: "${baseUrl}/api/gms"
-        # Add a secret in secrets Tab
-        token: "\${GMS_TOKEN}"`;
+`;
 
 const redshiftConfig: SourceConfig = {
     type: 'redshift',
