+++ metadata-ingestion/src/datahub/emitter/rest_emitter.py
@@ -1,8 +1,6 @@
 import datetime
-import itertools
 import json
 import logging
-import shlex
 from json.decoder import JSONDecodeError
 from typing import Any, Dict, List, Optional, Tuple, Union
 
@@ -10,8 +8,10 @@ import requests
 from requests.adapters import HTTPAdapter, Retry
 from requests.exceptions import HTTPError, RequestException
 
+from datahub.cli.cli_utils import get_system_auth
 from datahub.configuration.common import ConfigurationError, OperationalError
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
+from datahub.emitter.request_helper import _make_curl_command
 from datahub.emitter.serialization_helper import pre_json_transform
 from datahub.metadata.com.linkedin.pegasus2avro.mxe import (
     MetadataChangeEvent,
@@ -20,23 +20,6 @@ from datahub.metadata.com.linkedin.pegasus2avro.mxe import (
 from datahub.metadata.com.linkedin.pegasus2avro.usage import UsageAggregation
 
 logger = logging.getLogger(__name__)
-
-
-def _make_curl_command(
-    session: requests.Session, method: str, url: str, payload: str
-) -> str:
-    fragments: List[str] = [
-        "curl",
-        *itertools.chain(
-            *[
-                ("-X", method),
-                *[("-H", f"{k!s}: {v!s}") for (k, v) in session.headers.items()],
-                ("--data", payload),
-            ]
-        ),
-        url,
-    ]
-    return " ".join(shlex.quote(fragment) for fragment in fragments)
 
 
 class DataHubRestEmitter:
@@ -91,6 +74,10 @@ class DataHubRestEmitter:
         )
         if token:
             self._session.headers.update({"Authorization": f"Bearer {token}"})
+        else:
+            system_auth = get_system_auth()
+            if system_auth is not None:
+                self._session.headers.update({"Authorization": system_auth})
 
         if extra_headers:
             self._session.headers.update(extra_headers)
