+++ metadata-service/factories/src/main/resources/application.yml
@@ -42,7 +42,7 @@ authorization:
 
 ingestion:
   enabled: ${UI_INGESTION_ENABLED:true}
-  defaultCliVersion: '${UI_INGESTION_DEFAULT_CLI_VERSION:0.8.38}'
+  defaultCliVersion: '${UI_INGESTION_DEFAULT_CLI_VERSION:0.8.38.2}'
 
 telemetry:
   enabledCli: ${CLI_TELEMETRY_ENABLED:true}
